Created a constructor for deep copying
Created a constructor that will take any String object and perform a deep copy.
Achieved using a for loop.

Below the destructor in int main() there's an example of its usage (testing shows that it will work no matter the length of the string. (This example will be removed on my next commit)