Added ToLower() + Bug Fixes
Added a function to make all letters in the string into a lowercase.

Also added an example of its usage which will be removed or changed by next commit (as per usual.)

Also, maybe a bug fix, which is still kind of bugged, related to the append function, specifically changing tempLength into str.lengthOfString so it doesn't go way over. Will fix this more later.