Append tester added + BUG FIXES
+ Added:
Added a tester for the append function.

~ Bug Fix:
Fixed a bug related to memory in the append function. Related to the length of string which didn't change even after the string had another string appended to it. Fixed it, so now it should also get the memory of the new appended string. Go me!